+++ api/mathsys/compile.py
@@ -17,7 +17,8 @@ router = APIRouter()
 # FUNCTION -> EXTENSIONS
 from website.extensions import (
     mathsys as _mathsys,
-    post as _post
+    post as _post,
+    response as _response
 )
 
 # FUNCTION -> DECLARATION
@@ -26,10 +27,14 @@ async def output(request: Request) -> JSONResponse:
     # DECLARATION -> ACTIVATE EXTENSIONS
     mathsys = await _mathsys.namespace().init(request)
     post = await _post.namespace().init(request)
+    response = await _response.namespace().init(request)
     # DECLARATION -> ARGUMENT CHECKS
     if not post.checks(): raise HTTPException(**SUG.ERR[0])
     # DECLARATION -> ARGUMENT RELATIONSHIP
     if not post.exists("Mcode"): raise HTTPException(**SUG.ERR[1])
     # DECLARATION -> QUERY
-    try: return JSONResponse(content = mathsys.compile(post.data["Mcode"]))
-    except: return JSONResponse(content = False)
+    try:
+        response.load(mathsys.compile(post.data["Mcode"]))
+    except:
+        response.load(False)
+    return response.get()