+++ api/problem/lookup.py
@@ -17,8 +17,8 @@ router = APIRouter()
 from website.extensions import (
     binary as _binary,
     database as _database,
-    json as _json,
-    post as _post
+    post as _post,
+    response as _response
 )
 
 # FUNCTION -> DECLARATION
@@ -27,18 +27,18 @@ async def output(request: Request) -> JSONResponse:
     # DECLARATION -> ACTIVATE EXTENSIONS
     binary = await _binary.namespace().init(request)
     database = await _database.namespace().init(request)
-    json = await _json.namespace().init(request)
     post = await _post.namespace().init(request)
+    response = await _response.namespace().init(request)
     # DECLARATION -> ARGUMENT CHECKS
     if not post.checks(): raise HTTPException(**SUG.ERR[0])
     # DECLARATION -> ARGUMENT RELATIONSHIP
     if not post.exists("Pid"): raise HTTPException(**SUG.ERR[1])
     # DECLARATION -> QUERY
-    json.load((await database.query(
+    response.load(º(await database.query(
         "SELECT * FROM PROBLEMS WHERE Pid = %s",
         [
             binary.str2bin(post.data["Pid"])
         ]
-    ))[0])
-    json.tojson("Pmeta", "Pdataen", "Pdataes", "Pdatade", "Psolution")
-    return JSONResponse(content = json.data)
+    ), 0))
+    response.tojson("Pmeta", "Pdataen", "Pdataes", "Pdatade", "Psolution")
+    return response.get()