+++ package/update.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd /srv/www/website/database
+mysql -u root -p < update.sql